refactor: code structure (#45)
* refactor: code structure

changes:

- renaming some packages name.
- move reusable package to /pkg
- refactor utilities package like printer and id generator
- eliminating non efficient initialization of package.
- leave some TODO comments
- fix capitalized error message

* fix: handle octavius log initialization error
