+++ cmd/cli/main.go
@@ -1,19 +1,21 @@
 package main
 
 import (
+	"log"
 	"octavius/internal/cli/command"
 	"octavius/internal/cli/config"
 	"octavius/internal/cli/daemon"
-	"octavius/internal/cli/fileUtil"
-	"octavius/internal/cli/logger"
-	"octavius/internal/cli/printer"
+	octlog "octavius/internal/pkg/log"
 )
 
 func main() {
-	newPrinter := printer.GetPrinter()
-	logger.Setup()
-	fileUtil := fileUtil.NewFileUtil()
+	// TODO: get log level from config and define log file path if any
+	if err := octlog.Init("info", ""); err != nil {
+		log.Fatal("fail to initialize octavius log")
+	}
+
 	clientConfigLoader := config.NewLoader()
 	octaviusDaemon := daemon.NewClient(clientConfigLoader)
-	command.Execute(octaviusDaemon, fileUtil, newPrinter)
+
+	command.Execute(octaviusDaemon)
 }
