--- cmd/control_plane/main.go
@@ -1,13 +1,0 @@
-package main
-
-import (
-	"octavius/internal/control_plane/command"
-	"octavius/internal/control_plane/config"
-	"octavius/internal/control_plane/logger"
-)
-
-func main() {
-	logLevel := config.Config().LogLevel
-	logger.Setup(logLevel)
-	command.Execute()
-}
