+++ go.mod
@@ -7,8 +7,6 @@ require (
 	github.com/fatih/color v1.7.0
 	github.com/gogo/protobuf v1.2.1
 	github.com/golang/protobuf v1.4.1
-	github.com/mitchellh/go-homedir v1.1.0
-	github.com/prometheus/common v0.4.0
 	github.com/rs/zerolog v1.19.0
 	github.com/sony/sonyflake v1.0.0
 	github.com/spf13/cobra v1.0.0
