+++ internal/cli/command/create/create.go
@@ -5,12 +5,12 @@ import (
 	"github.com/spf13/cobra"
 	"octavius/internal/cli/client"
 	"octavius/internal/cli/daemon"
-	"octavius/internal/cli/fileUtil"
-	"octavius/internal/cli/logger"
+	"octavius/internal/pkg/file"
+	"octavius/internal/pkg/log"
 )
 
 // NewCmd Returns an instance of Create command for registering Job Metadata in Octavius
-func NewCmd(octaviusDaemon daemon.Client, fileUtil fileUtil.FileUtil) *cobra.Command {
+func NewCmd(octaviusDaemon daemon.Client, fileUtil file.File) *cobra.Command {
 	var metadataFilePath string
 
 	createCmd := &cobra.Command{
@@ -20,19 +20,21 @@ func NewCmd(octaviusDaemon daemon.Client, fileUtil fileUtil.FileUtil) *cobra.Com
 		Example: fmt.Sprintf("octavius create --job-path <filepath>/metadata.json"),
 
 		Run: func(cmd *cobra.Command, args []string) {
+
 			metadataFileIoReader, err := fileUtil.GetIoReader(metadataFilePath)
 			if err != nil {
-				logger.Error(err, "")
+				log.Error(err, "")
 				return
 			}
 
 			client := &client.GrpcClient{}
 			res, err := octaviusDaemon.CreateMetadata(metadataFileIoReader, client)
 			if err != nil {
-				logger.Error(err, "")
+				log.Error(err, "")
 				return
 			}
-			logger.Info(fmt.Sprintf("%s job created", res.Name))
+			
+			log.Info(fmt.Sprintf("%s job created", res.Name))
 		},
 	}
 	createCmd.Flags().StringVarP(&metadataFilePath, "job-path", "", "", "path to metadata.json(required)")
