+++ internal/cli/command/create/create_test.go
@@ -3,8 +3,8 @@ package create
 import (
 	"errors"
 	"octavius/internal/cli/daemon"
-	"octavius/internal/cli/fileUtil"
-	"octavius/internal/cli/logger"
+	"octavius/internal/pkg/file"
+	"octavius/internal/pkg/log"
 	protobuf "octavius/internal/pkg/protofiles/client_CP"
 	"strings"
 	"testing"
@@ -13,12 +13,12 @@ import (
 )
 
 func init() {
-	logger.Setup()
+	log.Init("info", "")
 }
 
 func TestCreateCmdHelp(t *testing.T) {
 	mockOctaviusDClient := new(daemon.MockClient)
-	mockFileUtil := new(fileUtil.MockFileUtil)
+	mockFileUtil := new(file.MockFileUtil)
 	testCreateCmd := NewCmd(mockOctaviusDClient, mockFileUtil)
 	assert.Equal(t, "Create new octavius job metadata", testCreateCmd.Short)
 	assert.Equal(t, "This command helps create new jobmetadata to your CP host with proper metadata.json file", testCreateCmd.Long)
@@ -27,7 +27,7 @@ func TestCreateCmdHelp(t *testing.T) {
 
 func TestCreateCmd(t *testing.T) {
 	mockOctaviusDClient := new(daemon.MockClient)
-	mockFileUtil := new(fileUtil.MockFileUtil)
+	mockFileUtil := new(file.MockFileUtil)
 	testCreateCmd := NewCmd(mockOctaviusDClient, mockFileUtil)
 	testMetadataName := &protobuf.MetadataName{
 		Name: "name",
@@ -44,7 +44,7 @@ func TestCreateCmd(t *testing.T) {
 
 func TestCreateCmdForIoError(t *testing.T) {
 	mockOctaviusDClient := new(daemon.MockClient)
-	mockFileUtil := new(fileUtil.MockFileUtil)
+	mockFileUtil := new(file.MockFileUtil)
 	testCreateCmd := NewCmd(mockOctaviusDClient, mockFileUtil)
 	mockFileUtil.On("GetIoReader", "testfile/test_metadata.json").Return(strings.NewReader(""), errors.New("test io error"))
