+++ internal/cli/command/execution/execution_test.go
@@ -1,15 +1,16 @@
 package execution
 
 import (
-	"github.com/stretchr/testify/assert"
 	"octavius/internal/cli/daemon"
-	"octavius/internal/cli/logger"
+	"octavius/internal/pkg/log"
 	protobuf "octavius/internal/pkg/protofiles/client_CP"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func init() {
-	logger.Setup()
+	log.Init("info", "")
 }
 
 func TestExecuteCmdHelp(t *testing.T) {
