+++ internal/controller/server/execution/execution.go
@@ -1,0 +1,36 @@
+package execution
+
+import (
+	"context"
+	"octavius/internal/controller/server/metadata/repository"
+	protobuf "octavius/internal/pkg/protofiles/client_CP"
+)
+
+// Execution interface for methods related to execution
+type Execution interface {
+	SaveMetadataToDb(ctx context.Context, metadata *protobuf.Metadata) (*protobuf.MetadataName, error)
+	ReadAllMetadata(ctx context.Context) (*protobuf.MetadataArray, error)
+}
+
+type execution struct {
+	metadata repository.MetadataRepository
+}
+
+// NewExec creates a new instance of metadata respository
+func NewExec(metadataRepo repository.MetadataRepository) Execution {
+	return &execution{
+		metadata: metadataRepo,
+	}
+}
+
+//SaveMetadataToDb calls the metadata/repository Save() function and returns MetadataName
+func (e *execution) SaveMetadataToDb(ctx context.Context, metadata *protobuf.Metadata) (*protobuf.MetadataName, error) {
+	result, err := e.metadata.Save(ctx, metadata.Name, metadata)
+	return result, err
+}
+
+//ReadAllMetadata calls the metadata/repository GetAll() and returns MetadataArray
+func (e *execution) ReadAllMetadata(ctx context.Context) (*protobuf.MetadataArray, error) {
+	result, err := e.metadata.GetAll(ctx)
+	return result, err
+}
